requirements.txt
absl-py==1.0.0; python_version >= "3.6"
astunparse==1.6.3
black==21.11b1; python_full_version >= "3.6.2"
cached-property==1.5.2; python_version < "3.8" and python_version >= "3.7"
cachetools==4.2.4; python_version >= "3.5" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
charset-normalizer==2.0.8; python_full_version >= "3.6.0" and python_version >= "3.6"
click==8.0.3; python_version >= "3.6" and python_full_version >= "3.6.2"
colorama==0.4.4; python_version >= "3.6" and python_full_version >= "3.6.2" and platform_system == "Windows"
cycler==0.11.0; python_version >= "3.7"
easydict==1.9
flatbuffers==2.0
fonttools==4.28.2; python_version >= "3.7"
gast==0.4.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
google-auth-oauthlib==0.4.6; python_version >= "3.6"
google-auth==2.3.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
google-pasta==0.2.0
grpcio==1.42.0; python_version >= "3.6"
h5py==3.6.0; python_version >= "3.7"
idna==3.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
imageio==2.13.1; python_version >= "3.7"
importlib-metadata==4.8.2; python_version >= "3.6" and python_full_version >= "3.6.2" and python_version < "3.8"
keras-preprocessing==1.1.2
keras==2.7.0
kiwisolver==1.3.2; python_version >= "3.7"
libclang==12.0.0
llvmlite==0.33.0; python_version >= "3.6"
mako==1.1.6; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
markdown==3.3.6; python_version >= "3.6"
markupsafe==2.0.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
matplotlib==3.5.0; python_version >= "3.7"
mnist @ file:///home/tuananhroman/DAI/constrastive-explaination-prototype/experiments/mnist; python_version >= "3.6" and python_version < "4.0"
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
networkx==2.6.3; python_version >= "3.7"
nptyping==1.4.4; python_version >= "3.5"
numba==0.50.0; python_version >= "3.6"
numpy==1.21.4; python_version >= "3.7" and python_version < "3.11"
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
opencv-python==4.5.4.60; python_version >= "3.6"
opt-einsum==3.3.0; python_version >= "3.5"
packaging==21.3; python_version >= "3.7"
pathspec==0.9.0; python_full_version >= "3.6.2"
pdoc3==0.10.0; python_version >= "3.6"
pillow==8.4.0; python_version >= "3.7"
platformdirs==2.4.0; python_version >= "3.6" and python_full_version >= "3.6.2"
protobuf==3.19.1; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
pyparsing==3.0.6; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
pywavelets==1.2.0; python_version >= "3.7"
pyyaml==5.4.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
regex==2021.11.10; python_full_version >= "3.6.2"
requests-oauthlib==1.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.26.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
scikit-image==0.18.3; python_version >= "3.7"
scipy==1.7.3; python_version >= "3.7" and python_version < "3.11"
setuptools-scm==6.3.2; python_version >= "3.7"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
tensorboard-data-server==0.6.1; python_version >= "3.6"
tensorboard-plugin-wit==1.8.0; python_version >= "3.6"
tensorboard==2.7.0; python_version >= "3.6"
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0; python_version >= "3.6" and python_version < "3.10"
termcolor==1.1.0
tifffile==2021.11.2; python_version >= "3.7"
tomli==1.2.2; python_version >= "3.7" and python_full_version >= "3.6.2"
torch==1.10.0; python_full_version >= "3.6.2"
typed-ast==1.5.0; python_version < "3.8" and implementation_name == "cpython" and python_full_version >= "3.6.2" and python_version >= "3.6"
typing-extensions==4.0.1; python_version >= "3.6" and python_full_version >= "3.6.2" and python_version < "3.8"
typish==1.9.3; python_version >= "3.5"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
werkzeug==2.0.2; python_version >= "3.6"
wrapt==1.13.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
zipp==3.6.0; python_version >= "3.6" and python_full_version >= "3.6.2" and python_version < "3.8"
